[WIP] 🚧 Fix: [UI] BottomSheet
Improvements and PR fixes for [UI] BottomSheet
---
- Added separate file for WizardDataSource.swift
- Changed year in Copyright header
- Added private extension in WizardContentView
