+++ Sources/UI/Controllers/Wizard/Wizard.swift
@@ -3,50 +3,13 @@
 //  PovioKit
 //
 //  Created by Toni Kocjan on 29/09/2021.
-//  Copyright © 2021 Povio Labs. All rights reserved.
+//  Copyright © 2022 Povio Labs. All rights reserved.
 //
 
 import Foundation
 import UIKit
 
-public protocol WizardStep: UIViewController {
-}
-
-public class WizardDataSource: NSObject {
-  public typealias LazyStep = () -> WizardStep
-  
-  var currentStepIndex = -1
-  var currentStep: WizardStep?
-  var steps: [LazyStep]
-  
-  init(steps: [LazyStep]) {
-    self.steps = steps
-  }
-}
-
-public extension WizardDataSource {
-  func nextStep() -> WizardStep? {
-    currentStepIndex += 1
-    guard let thunk = steps[safe: currentStepIndex] else {
-      currentStepIndex -= 1
-      return nil
-    }
-    let step = thunk()
-    self.currentStep = step
-    return step
-  }
-  
-  func previousStep() -> WizardStep? {
-    currentStepIndex -= 1
-    guard let thunk = steps[safe: currentStepIndex] else {
-      currentStepIndex += 1
-      return nil
-    }
-    let step = thunk()
-    self.currentStep = step
-    return step
-  }
-}
+public protocol WizardStep: UIViewController { }
 
 /// An automatically resizable dialog superclass.
 ///
