+++ Sources/UI/Controllers/Wizard/WizardAnimators.swift
@@ -3,7 +3,7 @@
 //  PovioKit
 //
 //  Created by Toni Kocjan on 29/09/2021.
-//  Copyright © 2021 Povio Labs. All rights reserved.
+//  Copyright © 2022 Povio Labs. All rights reserved.
 //
 
 import UIKit
